+++ escrow/project/contracts/escrow-contract/tests/functions/info/arbiter_proposal.rs
@@ -1,18 +1,21 @@
 mod success {
 
     use crate::utils::{
-        abi_calls::{assets, create_escrow},
-        test_helpers::{create_arbiter, create_asset, mint, setup},
+        interface::{
+            core::{create_escrow, propose_arbiter},
+            info::arbiter_proposal,
+        },
+        setup::{create_arbiter, create_asset, mint, setup},
     };
 
     #[tokio::test]
     async fn returns_none() {
         let (_arbiter, _buyer, seller, _defaults) = setup().await;
-        assert!(matches!(assets(&seller.contract, 0).await, None));
+        assert!(matches!(arbiter_proposal(&seller.contract, 0).await, None));
     }
 
     #[tokio::test]
-    async fn returns_asset() {
+    async fn return_arbiter() {
         let (arbiter, buyer, seller, defaults) = setup().await;
         let arbiter_obj = create_arbiter(
             arbiter.wallet.address(),
@@ -24,24 +27,28 @@ mod success {
 
         mint(
             seller.wallet.address(),
-            defaults.asset_amount,
+            defaults.asset_amount * 2,
             &defaults.asset,
         )
         .await;
-
-        assert!(matches!(assets(&seller.contract, 0).await, None));
-
         create_escrow(
             defaults.asset_amount,
             &arbiter_obj,
             &defaults.asset_id,
-            vec![asset.clone()],
+            vec![asset.clone(), asset.clone()],
             buyer.wallet.address(),
             &seller.contract,
             defaults.deadline,
         )
         .await;
 
-        assert_eq!(assets(&seller.contract, 0).await.unwrap(), asset);
+        assert!(matches!(arbiter_proposal(&seller.contract, 0).await, None));
+
+        propose_arbiter(arbiter_obj.clone(), &seller.contract, 0).await;
+
+        assert_eq!(
+            arbiter_proposal(&seller.contract, 0).await.unwrap(),
+            arbiter_obj
+        );
     }
 }
